Added blocking implementation (#18)
This is a small PR to allow consumers to use it in non-async projects.

One thing I am not that happy about is the tests. The test module just
doubled in size as I had to copy/paste every async test for their
blocking counterpart. I thought about splitting the test module up into
"async_tests" and "blocking_tests" for clarity but perhaps I'm
overthinking it.